+++ fail/fail.go
@@ -21,8 +21,8 @@ const (
 	ErrSlotNotDefined         = "@slot('%s') not defined in @component('%s')"
 	ErrDuplicateSlot          = "@slot('%s') used %d times in @component('%s')"
 	ErrDuplicateDefaultSlot   = "default @slot used %d times in @component('%s')"
-	ErrExpectedComponentName  = "@component() cannot have empty name"
-	ErrUndefinedInsert        = "@insert('%s') not found in layout - add matching @reserve"
+	ErrExpectedComponentName  = "@component('') cannot have empty name"
+	ErrAddMatchingReserve     = "@insert('%s') needs to have a matching @reserve('%s') in layout file"
 	ErrDuplicateInserts       = "duplicate @insert('%s') found"
 	ErrUseStmtFirstArgStr     = "argument 1 of @use(STR) must be a string, got @use('%s')"
 	ErrOnlyOneUseDir          = "@use() directive can only be used once per template"
@@ -36,13 +36,14 @@ const (
 	ErrTypeMismatch            = "type mismatch: cannot %s %s %s"
 	ErrUnknownTypeForOp        = "operator '%s' not supported for type '%s'"
 	ErrPrefixOpIsWrong         = "cannot apply prefix '%s' to type '%s'"
-	ErrIdentifierIsUndefined   = "variable '%s' is not defined"
+	ErrVariableIsUndefined     = "variable '%s' is not defined"
 	ErrReservedIdentifiers     = "'loop' and 'global' are reserved variable names"
 	ErrIdentifierTypeMismatch  = "cannot assign identifier '%s' of type '%s' to type '%s'"
 	ErrPropertyNotFound        = "property '%s' not found on type '%s'"
 	ErrDivisionByZero          = "division by zero - divisor cannot be zero"
 	ErrEachDirWithNonArrArg    = "cannot use @each(item in ARRAY) with non-array type '%s' after 'in' keyword"
 	ErrSomeDirsOnlyInTemplates = "@use, @insert, @reserve, @component only allowed in templates"
+	ErrInsertRequiresUse       = "@insert('%s') cannot be used without @use()"
 	ErrUseStmtMissingLayout    = "@use('%s') missing layout file"
 	ErrGlobalFuncMissing       = "global function %s() not found"
 	ErrPropertyOnNonObject     = "'%s' type does not support attribute '%s' access"
@@ -88,6 +89,10 @@ func New(line uint, filepath, origin, msg string, args ...any) *Error {
 
 func (e *Error) Filepath() string {
 	return e.filepath
+}
+
+func (e *Error) Origin() string {
+	return e.origin
 }
 
 func (e *Error) Line() uint {