+++ object/reserve.go
@@ -1,5 +1,7 @@
 package object
 
+import "fmt"
+
 type Reserve struct {
 	Name   string
 	Insert Object
@@ -13,19 +15,17 @@ func (r *Reserve) String() string {
 	if r.Insert == nil {
 		panic("Insert field on Reseve object must not be nil when calling String()")
 	}
-
 	return r.Insert.String()
 }
 
 func (r *Reserve) Dump(ident int) string {
-	return "reserve stmt"
+	return fmt.Sprintf("@reseve(%q)", r.Name)
 }
 
 func (r *Reserve) Val() any {
 	if r.Insert == nil {
 		panic("Insert field on Reseve object must not be nil when calling Val()")
 	}
-
 	return r.Insert.Val()
 }